clippy

+++ core/server/src/eth_sender/mod.rs
@@ -692,11 +692,9 @@ impl<ETH: EthereumInterface, DB: DatabaseAccess> ETHSender<ETH, DB> {
                     (
                         u64::from(op.block.block_number),
                         u64::from(op.block.fee_account),
-                        u64::from(
-                            *op.block
-                                .block_timestamp
-                                .expect("block timestamp should be known at this moment"),
-                        ),
+                        *op.block
+                            .block_timestamp
+                            .expect("block timestamp should be known at this moment"),
                         vec![root],
                         public_data,
                         witness_data.0,
