+++ core/testkit/src/eth_account.rs
@@ -373,11 +373,9 @@ impl<T: Transport> EthereumAccount<T> {
                 (
                     u64::from(block.block_number),
                     u64::from(block.fee_account),
-                    u64::from(
-                        *block
-                            .block_timestamp
-                            .expect("block timestamp should be known at this moment"),
-                    ),
+                    *block
+                        .block_timestamp
+                        .expect("block timestamp should be known at this moment"),
                     vec![block.get_eth_encoded_root()],
                     block.get_eth_public_data(),
                     witness_data.0,
